+++ api/bro_upload/type_helpers.py
@@ -55,6 +55,8 @@ RegistrationTypeOptions = Literal[
     "GPD_EndRegistration",
 ]
 
+LimitSymbolOptions = Literal["LT", "GT"]
+
 # GUF-specific value lists (from Section 6 of the BRO GUF Catalog)
 
 # 6.1 Bodemlustype (Soil Loop Type)