+++ gar/choices.py
@@ -1,0 +1,6 @@
+QUALITY_CONTROL_STATUS = (
+    ("afgekeurd", "Afgekeurd"),
+    ("goedgekeurd", "Goedgekeurd"),
+    ("onbeslist", "Onbeslist"),
+    ("onbekend", "Onbekend"),
+)